sim_py/requirements.txt
numpy>=1.22
scipy>=1.10
matplotlib>=3.5
pyyaml>=6.0

